moved logic to components on create-scenario
